requirements.txt
discord.py
yt-dlp
PyNaCl
aiohttp
python-dotenv